+++ packages/luthor/lib/src/annotations/validators/cuid.dart
@@ -1,0 +1,8 @@
+class IsCuid {
+  final String? message;
+  final String? Function()? messageFn;
+
+  const IsCuid({this.message, this.messageFn});
+}
+
+const isCuid = IsCuid();