+++ packages/luthor/lib/src/annotations/validators/cuid2.dart
@@ -1,0 +1,8 @@
+class IsCuid2 {
+  final String? message;
+  final String? Function()? messageFn;
+
+  const IsCuid2({this.message, this.messageFn});
+}
+
+const isCuid2 = IsCuid2();