+++ packages/luthor/lib/src/annotations/validators/emoji.dart
@@ -1,0 +1,8 @@
+class IsEmoji {
+  final String? message;
+  final String? Function()? messageFn;
+
+  const IsEmoji({this.message, this.messageFn});
+}
+
+const isEmoji = IsEmoji();