+++ packages/luthor/lib/src/annotations/validators/uuid.dart
@@ -1,0 +1,8 @@
+class IsUuid {
+  final String? message;
+  final String? Function()? messageFn;
+
+  const IsUuid({this.message, this.messageFn});
+}
+
+const isUuid = IsUuid();